requirements.txt
solders==0.18.1
solana==0.30.2
base58==2.1.1
python-dotenv==1.0.0
anchorpy==0.18.0 